fix: emit hash() as signed int32
The docs and the game interpret CRC-32 device/name hashes as signed
32-bit integers (e.g. hash("StructureBattery") == -400115994), but the
compiler folded hash() to the unsigned uint32 value (3894851302), which
overflows an int32 when the game reads it. Fold to int32 instead; the VM
already normalises either form. Update the m3 golden and add a test.

+++ internal/lower/lower.go
@@ -839,7 +839,7 @@ func (l *lowerer) lowerCallExpr(e ast.Expr, needResult bool) ir.Value {
 			l.diags.Errorf(call.Args[0].Pos(), "hash expects a string literal")
 			return &ir.Const{V: 0}
 		}
-		return &ir.Const{V: float64(builtin.Hash(s.Value))}
+		return &ir.Const{V: float64(int32(builtin.Hash(s.Value)))}
 	}
 
 	// str("...") produces a display string hash.
